1
1

+++ src/main/webapp/WEB-INF/views/qna/qnaContents.jsp
@@ -15,15 +15,56 @@
 <script type="text/javascript">
 	$(function() {
 			var num=${contents.boardNum}
+			var admin="${admin}";
+		
+			if(admin=="관리자"){
+				$("#delete").click(function(){
+					alert("삭제되었습니다.");
+					location.href="./qnaDelete.do?boardNum="+num;
+				});
+				
+				$("#reply").click(function(){
+					location.href="./qnaAdminForm.do?boardNum="+num;			
+				});
+			}else{
+			
 		$("#update").click(function() {
-			location.href="./qnaUpdate.do?boardNum="+num;
+			$("#deleted").html("수정");
+			$("#myModal").css("display","block");
 		});
 		$("#delete").click(function(){
-			location.href="./qnaDelete.do?boardNum="+num;			
+			/* location.href="./communityDelete.do?boardNum="+num;	 */	
+			$("#myModal").css("display","block");
+			$("#deleted").html("삭제");
 		});
+		$("#close").click(function(){
+			/* location.href="./communityDelete.do?boardNum="+num;	 */	
+			$("#myModal").css("display","none");
+		});
+		
 		$("#reply").click(function(){
-			location.href="./qnaAdminForm.do?boardNum="+num;			
+			alert("관리자만 답글 가능합니다.");			
 		});
+		
+		$(".updated").click(function(){
+			var pw= $("#boardPw").val()+"";
+			var boardPw="${contents.boardPw}";
+			var pageUrl=$("#deleted").text();
+			if(pw == boardPw){
+				if(pageUrl=="수정"){
+					location.href="./qnaUpdate.do?boardNum="+num;
+				}else{
+					alert("삭제되었습니다.");
+					location.href="./qnaDelete.do?boardNum="+num;
+				}
+			}else{
+				alert("비밀번호가 다릅니다.");
+			}
+		});
+		
+	
+			}
+			
 	});
 </script>
 </head>
@@ -50,6 +91,16 @@
 		</div>
 		<div id="reply">답변 달기 </div>
 	</section>
+
+<div id="myModal" class="modal">
+  <!-- Modal content -->
+  <div class="modal-content">
+    <span id="close">&times;</span>
+    <p> ▶ 비밀번호</p>
+    <p><input type="password" id="boardPw" name="boardPw" placeholder="비밀번호를 입력해주세요"></p>
+    <div class="writeButton updated" id="deleted">삭제</div>
+  </div>
+</div>
 
 </body>
 </html>
